+++ src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import './App.scss';
-import Wrapper from './components/Wrapper';
+import Wrapper from './components/Wrapper/Wrapper';
 import Footer from './components/Footer/Footer';
 
 function App() {
