+++ src/components/Footer/Footer.scss
@@ -1,3 +1,5 @@
+@import "../../styles/index";
+
 .footer {
   height: 540px;
   width: 100vw;
