+++ src/components/Hero/Hero.scss
@@ -1,4 +1,4 @@
-@import "../../index";
+@import "../../styles/index";
 
 .hero {
   height: 100vh;
