+++ src/components/SectionBest/SectionBest.scss
@@ -1,3 +1,5 @@
+@import "../../styles/index";
+
 .best {
   &__title {
     font-weight: 300;
