+++ src/components/SectionConsultation/SectionConsultation.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import './SectionConsultation.scss';
 
 const SectionConsultation = () => {
   return (
