+++ src/components/SectionConsultation/SectionConsultation.scss
@@ -1,3 +1,5 @@
+@import "../../styles/index";
+
 .container--consultation {
   display: flex;
   justify-content: space-between;
@@ -8,3 +10,4 @@
     flex-direction: row;
   }
 }
+
