+++ src/components/SectionContact/SectionContact.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import './SectionContact.scss';
 
 const SectionContact = () => {
   return (
