+++ src/components/SectionContact/SectionContact.scss
@@ -1,3 +1,5 @@
+@import "../../styles/index";
+
 .container--contact {
   display: flex;
   justify-content: space-between;
