+++ src/components/SectionGallery/SectionGallery.scss
@@ -1,3 +1,5 @@
+@import "../../styles/index";
+
 .gallery {
   text-align: center;
 }
