+++ src/components/SectionTestimonials/SectionTestimonials.scss
@@ -1,3 +1,5 @@
+@import "../../styles/index";
+
 .testimonials {
   &__grid {
     display: grid;
