--- src/components/Wrapper.js
@@ -1,22 +1,0 @@
-import React from 'react';
-import Hero from './Hero/Hero';
-import SectionBest from './SectionBest';
-import SectionGallery from './SectionGallery';
-import SectionConsultation from './SectionConsultation';
-import SectionTestimonials from './SectionTestimonials';
-import SectionContact from "./SectionContact";
-
-const Wrapper = () => {
-  return (
-    <div className="wrapper">
-      <Hero />
-      <SectionBest />
-      <SectionGallery />
-      <SectionConsultation />
-      <SectionTestimonials />
-      <SectionContact />
-    </div>
-  );
-};
-
-export default Wrapper;
