--- src/index.scss
@@ -1,41 +1,0 @@
-$media-desktop: 1024px;
-$media-tablet: 768px;
-$media-phone: 320px;
-
-$hover-duration: 300ms;
-
-$color-hover: #00aaef;
-$color-button-hover: #0076a7;
-
-$input-heigth: 54px;
-
-@mixin for-phone {
-  @media only screen and (max-width: #{$media-tablet - 1}) {
-    @content;
-  }
-}
-
-@mixin for-tablet {
-  @media only screen and (min-width: #{$media-tablet}) {
-    @content;
-  }
-}
-
-@mixin for-desktop {
-  @media only screen and (min-width: #{$media-desktop}) {
-    @content;
-  }
-}
-
-@mixin transition($duration, $property) {
-  transition-duration: $duration;
-  transition-property: #{$property};
-}
-
-@mixin transition-hover($duration, $property) {
-  @include transition ($duration, $property);
-
-  &:hover {
-    @content;
-  }
-}
