+++ src/styles/_blocks.scss
@@ -1,7 +1,2 @@
 @import "blocks/nav";
-@import "blocks/section_best";
-@import "blocks/section_gallery";
-@import "blocks/section_consultation";
 @import "blocks/form";
-@import "blocks/section_contact";
-@import "blocks/section_testimonials";
