--- src/styles/_layout.scss
@@ -1,2 +1,0 @@
-@import "layout/page";
-@import "layout/section";
