+++ src/styles/index.scss
@@ -1,0 +1,5 @@
+@import "utils/vars";
+@import "utils/mixins";
+
+@import "layout/section";
+@import "layout/page";
