+++ src/styles/main.scss
@@ -1,5 +1,4 @@
 @import "utils";
 @import "fonts";
 @import "typography";
-@import "layout";
 @import "blocks";
